% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp_Factors.R
\name{Decomp_Factors}
\alias{Decomp_Factors}
\title{Compute the marginal decomposition effects from given vectors}
\usage{
Decomp_Factors(vec_x, vec_y, return_dt = TRUE, ...)
}
\arguments{
\item{vec_x}{First input vector (represents t-1)}

\item{vec_y}{Second input vector (represents t)}

\item{return_dt}{A boolean on whether to return a data.table or a vector}

\item{...}{extra parameters to be passed to \code{all.equal()}}
}
\value{
A data.table or vector of decomposed effects for each factors,
which is already multiplied by the change values
}
\description{
Compute the marginal decomposition effects from given vectors
}
