% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp_on_DT.R
\name{Decomp_on_DT}
\alias{Decomp_on_DT}
\title{Apply DG Decomposition to data.table columns}
\usage{
Decomp_on_DT(input_data, factor_names, time_col, bycol, use_cpp = TRUE,
  ...)
}
\arguments{
\item{input_data}{A data.table with the factors, which must already
be sorted}

\item{factor_names}{A vector of column names for the factor}

\item{time_col}{A string for the column name}

\item{bycol}{The 'by' slicer which must make sure that the data is
reduced to just 2 rows per group after slicing}

\item{use_cpp}{A boolean on whether to use the C++ compiled code for the
factor for-loop or not (passes to Decomp_Factor_Matx()). Default: TRUE}

\item{...}{extra parameters to be passed through \code{Decomp_Factors()}
to \code{all.equal()}}
}
\value{
A data.table of the same size as input, but instead with
the additive decomposition results (first row will be NA as being the
starting period)
}
\description{
Apply DG Decomposition to data.table columns
}
