% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Create_Combn.R
\name{Func_Create_Combn}
\alias{Func_Create_Combn}
\title{Compute the combination of all positions of the given factor
segmented into two pieces}
\usage{
Func_Create_Combn(Pfac, size1, size2)
}
\arguments{
\item{Pfac}{Number of factors minus 1}

\item{size1}{Number of elements to take from vec_x}

\item{size2}{Number of elements to take from vec_y}
}
\value{
A vector of positions made from the unique combinations
of size1 and size2
}
\description{
Compute the combination of all positions of the given factor
segmented into two pieces
}
