% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Dem.R
\name{Func_Dem}
\alias{Func_Dem}
\title{Compute the denomiator of the inner sums in the DG Decomposition}
\usage{
Func_Dem(P, r)
}
\arguments{
\item{P}{Number of factors}

\item{r}{The summing indicator}
}
\value{
A numeric with value of \code{P * choose((P - 1), (r - 1))}
}
\description{
Compute the denomiator of the inner sums in the DG Decomposition
}
