% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Inner_Sum.R
\name{Func_Inner_Sum}
\alias{Func_Inner_Sum}
\title{Sum over all inner sums for the DG Decomposition}
\usage{
Func_Inner_Sum(P, vec_x, vec_y)
}
\arguments{
\item{P}{Number of factors}

\item{vec_x}{First input vector}

\item{vec_y}{Second input vector}
}
\value{
A numeric value with the full inner sum for the given effect
}
\description{
Sum over all inner sums for the DG Decomposition
}
