% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Num.R, R/Func_Num_Matx.R
\name{Func_Num}
\alias{Func_Num}
\alias{Func_Num_Matx}
\title{Compute the numerator of the inner sums in the DG Decomposition}
\usage{
Func_Num(P, r, vec_x, vec_y)

Func_Num_Matx(P, r, vec_x, vec_y)
}
\arguments{
\item{P}{Number of factors}

\item{r}{The summing indicator}

\item{vec_x}{First input vector}

\item{vec_y}{Second input vector}
}
\value{
A single numeric from the sums of Func_Cross()
}
\description{
Compute the numerator of the inner sums in the DG Decomposition
}
