% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_decomp_data.R
\name{simulate_decomp_data}
\alias{simulate_decomp_data}
\title{Simulate simple random decomp data for P factors such that the
product of P factors equal a measure for 2 time periods}
\usage{
simulate_decomp_data(num_fac)
}
\arguments{
\item{num_fac}{Number of factors to simulate}
}
\value{
A named list with the vector of P factors for 2 time periods,
and 2 numeric measures for each time period, which are just the
product of each of the two factor vectors
}
\description{
Simulate simple random decomp data for P factors such that the
product of P factors equal a measure for 2 time periods
}
