% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_decomp_data_fullmat.R
\name{simulate_decomp_data_fullmat}
\alias{simulate_decomp_data_fullmat}
\title{Simulate grouped data for decomp analysis (T by IDI by factors)}
\usage{
simulate_decomp_data_fullmat(T_term, num_factors, id_grps)
}
\arguments{
\item{T_term}{Number of time periods}

\item{num_factors}{Number of factors (gets slower for large numbers)}

\item{id_grps}{Number of groups}
}
\value{
A data.table with T_term*id_grps rows and num_factors+1 columns
where X_1, ... , X_p are the factors, and Y is the rowwise product of the
factors
}
\description{
Simulate grouped data for decomp analysis (T by IDI by factors)
}
