% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Cross.R
\name{Func_Cross}
\alias{Func_Cross}
\title{Compute the combination of all the elements of the given vectors
corresponding to the given sizes (using Func_Create_Combn)}
\usage{
Func_Cross(Pfac, vec_x, vec_y, size1, size2)
}
\arguments{
\item{Pfac}{Number of factors minus 1}

\item{vec_x}{First input vector}

\item{vec_y}{Second input vector}

\item{size1}{Number of elements to take from vec_x}

\item{size2}{Number of elements to take from vec_y}
}
\value{
A vector of column products made from the unique combinations
of the *data*
}
\description{
Compute the combination of all the elements of the given vectors
corresponding to the given sizes (using Func_Create_Combn)
}
