% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Inner.R
\name{Func_Inner}
\alias{Func_Inner}
\title{Compute the fraction for the inner sum in the DG Decomposition
where all the params gets passed to Func_Num() and Func_Den()}
\usage{
Func_Inner(P, r, vec_x, vec_y)
}
\arguments{
\item{P}{Number of factors}

\item{r}{The summing indicator}

\item{vec_x}{First input vector}

\item{vec_y}{Second input vector}
}
\value{
The fraction of the results of Func_Num() and Func_Den
}
\description{
Compute the fraction for the inner sum in the DG Decomposition
where all the params gets passed to Func_Num() and Func_Den()
}
